+++ dashboard/templates/common.html
@@ -18,6 +18,8 @@
 {% block javascripts %}
 
     <script src="https://cdn.plot.ly/plotly-1.10.2.min.js"></script>
+    <script src="https://code.jquery.com/jquery-2.2.3.js"></script>
+    <script src="{% static 'prototype/js/main.js' %}"></script>
 
 {% endblock %}
